// src/app/core/interfaces/Plantilla_Integracion.ts
export interface Plantilla_IntegracionI {
  pi_id: string;
  pi_codigo: string;
  pi_nombre: string;
  pi_descripcion: string;
  pi_ind_estado: string;
  pi_fecha_actividad: string;
  pi_tipo_servicio: string;
  pi_sist_orig_id: string;
  pi_data: string;
  pi_numreg_peticion: string;
  pi_schema: string;
  pi_url: string;
  pi_metodo_http: string;
  pi_sist_id: string;
  pi_cola_id: string;
  pi_valida: string;
  pi_transforma: string;
}